+++ test/step_test.rb
@@ -35,9 +35,9 @@ class StepTest < Minitest::Spec
 
   it { Create.({}, a: 1, b: 2, c: 3, d: 4, e: 5).inspect("a", "b", "c", "d", "e").must_equal "<Result:true [1, 2, 3, 4, 5] >" }
 
-  it { Trailblazer::Operation::Inspect.(Create).gsub(/0x[\w]+/, "").must_equal %{[>#<Proc:@test/step_test.rb:25 (lambda)>,>StepTest::Callable,>#<Method: StepTest::Implementation.c>,>d,>]} }
+  it { Trailblazer::Operation::Inspect.(Create).gsub(/0x.+?step_test.rb/, "").must_equal %{[>#<Proc::25 (lambda)>,>StepTest::Callable,>#<Method: StepTest::Implementation.c>,>d,>]} }
   # poor test to make sure we pass debug information to Activity.
-  it { Create["pipetree"].circuit.to_fields.last.to_a[3].last.must_equal :d }
+  it { Create["__activity__"].circuit.to_fields.last.to_a[3].last.must_equal :d }
 
   #---
   #- :before, :after, :replace, :delete, :override
@@ -69,6 +69,17 @@ class StepTest < Minitest::Spec
 
   it { Trailblazer::Operation::Inspect.(D).must_equal %{[>a!,>b!]} }
 
+  # not existent :name
+  it do
+    err = assert_raises Trailblazer::Operation::Railway::Sequence::IndexError  do
+      class E < Trailblazer::Operation
+        step :a, before: "I don't exist!"
+      end
+    end
+
+    err.inspect.must_equal "#<Trailblazer::Operation::Railway::Sequence::IndexError: I don't exist!>"
+  end
+
   #---
   #- :name
   #-   step :whatever, name: :validate
@@ -86,11 +97,11 @@ class StepTest < Minitest::Spec
   class New < Create
   end
 
-  it { Trailblazer::Operation::Inspect.(New).gsub(/0x[\w]+/, "").must_equal %{[>#<Proc:@test/step_test.rb:25 (lambda)>,>StepTest::Callable,>#<Method: StepTest::Implementation.c>,>d,>]} }
+  it { Trailblazer::Operation::Inspect.(New).gsub(/0x.+?step_test.rb/, "").must_equal %{[>#<Proc::25 (lambda)>,>StepTest::Callable,>#<Method: StepTest::Implementation.c>,>d,>]} }
 
   class Update < Create
     step :after_save!
   end
 
-  it { Trailblazer::Operation::Inspect.(Update).gsub(/0x[\w]+/, "").must_equal %{[>#<Proc:@test/step_test.rb:25 (lambda)>,>StepTest::Callable,>#<Method: StepTest::Implementation.c>,>d,>,>after_save!]} }
+  it { Trailblazer::Operation::Inspect.(Update).gsub(/0x.+?step_test.rb/, "").must_equal %{[>#<Proc::25 (lambda)>,>StepTest::Callable,>#<Method: StepTest::Implementation.c>,>d,>,>after_save!]} }
 end