+++ test/test_helper.rb
@@ -1,2 +1,6 @@
 require "minitest/autorun"
 require "trailblazer/operation"
+
+require "pp"
+
+require "raise"