+++ test/test_helper.rb
@@ -1,6 +1,6 @@
+require "raise"
+
 require "minitest/autorun"
 require "trailblazer/operation"
-
-require "pp"
 
 require "raise"