+++ lib/trailblazer/operation.rb
@@ -1,7 +1,7 @@
 require "forwardable"
 require "declarative"
 require "trailblazer/circuit"
-require "trailblazer/operation/public_call" # Remove in 3.0.
+require "trailblazer/operation/public_call"      # TODO: Remove in 3.0.
 require "trailblazer/operation/skill"
 require "trailblazer/operation/deprecated_macro" # TODO: remove in 2.2.
 require "trailblazer/operation/result"
@@ -28,10 +28,10 @@ module Trailblazer
     include Railway::TaskWrap
 
     # we want the skill dependency-mechanism.
-    extend Skill::Call             # ::call(params: .., current_user: ..)
+    # extend Skill::Call             # ::call(params: .., current_user: ..)
     extend PublicCall              # ::call(params, { current_user: .. })
 
-    extend Trace                   # ::trace
+    # extend Trace                   # ::trace
   end
 end