+++ lib/trailblazer/operation/public_call.rb
@@ -4,14 +4,20 @@ class Trailblazer::Operation
     # The signature of this is `params, options, *containers`. This was a mistake, as the
     # first argument could be part of `options` in the first place.
     #
+    # Create.(params, runtime_data, *containers)
+    #   #=> Result<skills...>
+    #
     # In workflows/Nested compositions, this method is not used anymore and it might probably
     # get removed in future versions of TRB.
     #
     # @returns Operation::Result binary result object
     def call(params={}, options={}, *containers)
-      options = options.merge("params" => params)
+      options = options.merge("params" => params) # options will be passed to all steps/activities.
 
-      direction, options, flow_options = super(options, *containers)   # FIXME: should we return the Result object here?
+      # generate the skill hash that embraces runtime options plus potential containers.
+      skills = Trailblazer::Skill.new(options, *containers, {})
+
+      direction, options, flow_options = super(skills) # DISCUSS: this could be ::call_with_skills.
 
       # Result is successful if the activity ended with the "right" End event.
       Railway::Result(direction, options)