+++ lib/trailblazer/operation/skill.rb
@@ -15,20 +15,21 @@ class Trailblazer::Operation
     end
 
     # Overrides Operation::call, creates the Skill hash and passes it to ::call.
-    module Call
-      def call(options={}, *dependencies)
-        super( Trailblazer::Operation::Skill(self, options, *dependencies) )
-      end
-    end # Call
+    # module Call
+    #   # def __call__(options={}, *dependencies)
+    #   def call(options={}, *dependencies)
+    #     super( Trailblazer::Operation::Skill(self, options, *dependencies) ) # FIXME: DO this in __call__ as we need the skill hash with self?
+    #   end
+    # end # Call
   end
 
   # Returns a `Skill` object that maintains all dependencies for this operation.
   # @returns Trailblazer::Skill
-  def self.Skill(operation, options, *dependencies)
-    Trailblazer::Skill.new(
-      options,
-      *dependencies,
-      operation.skills
-    )
-  end
+  # def self.Skill(operation, options, *dependencies) # DISCUSS: should this be a private concept?
+  #   Trailblazer::Skill.new(
+  #     options,
+  #     *dependencies,
+  #     operation.skills
+  #   )
+  # end
 end