+++ CHANGES.md
@@ -3,6 +3,18 @@
 * Remove `Uber::Callable` requirement and treat all non-`:symbol` steps as callable objects.
 * Remove support for Ruby 1.9.3 for now. This can be re-introduced on demand.
 * Remove non-kw options for steps. All steps receive keyword args now.
+
+    ```ruby
+    def model(options)
+    ```
+
+    now must have a minimal signature as follows.
+
+    ```ruby
+    def model(options, **)
+    ```
+
+
 * Remove `pipetree` and `uber` dependency.
 
 ## 0.0.13