+++ CHANGES.md
@@ -1,7 +1,9 @@
+TODO:
+* api to add your own task.
+
 ## 0.1.0
 
 * Remove `Uber::Callable` requirement and treat all non-`:symbol` steps as callable objects.
-* Remove support for Ruby 1.9.3 for now. This can be re-introduced on demand.
 * Remove non-kw options for steps. All steps receive keyword args now.
 
     ```ruby
@@ -14,7 +16,8 @@
     def model(options, **)
     ```
 
-
+* API change for `step Macro()`: the macro's return value is now called with the normal step signature `(options, kws**)` just like a normal step. There's a soft-deprecation warning.
+* Remove support for Ruby 1.9.3 for now. This can be re-introduced on demand.
 * Remove `pipetree` and `uber` dependency.
 
 ## 0.0.13